+++ database/migrations/2025_05_29_100139_add_phone_number_to_orders_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('orders', function (Blueprint $table) {
-            $table->string('phone_number')->nullable()->after('billing_email');
-        });
+        if (!Schema::hasColumn('orders', 'phone_number')) {
+            Schema::table('orders', function (Blueprint $table) {
+                $table->string('phone_number')->nullable()->after('billing_email');
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('orders', function (Blueprint $table) {
-            $table->dropColumn('phone_number');
-        });
+        if (Schema::hasColumn('orders', 'phone_number')) {
+            Schema::table('orders', function (Blueprint $table) {
+                $table->dropColumn('phone_number');
+            });
+        }
     }
 };
